+++ newIDE/app/src/ResourcesList/ResourcesManagementCallbacks.js
